+++ R/get_parameters.R
@@ -150,8 +150,8 @@ get_parameters.selection <- function(x, component = c("all", "selection", "outco
 
 #' @export
 get_parameters.epi.2by2 <- function(x, ...) {
-  coef_names <- grepl(".strata.wald", names(x$massoc), fixed = TRUE)
-  cf <- x$massoc[coef_names]
+  coef_names <- grepl(".strata.wald", names(x$massoc.detail), fixed = TRUE)
+  cf <- x$massoc.detail[coef_names]
   names(cf) <- gsub(".strata.wald", "", names(cf), fixed = TRUE)
 
   params <- data.frame(