+++ NEWS.md
@@ -1,5 +1,10 @@
 # insight 0.14.2
 
+## Changes to functions
+
+* `check_if_installed()` gains a `minimum_version` argument, to check if an
+  installed package is not older than the specified version number.
+
 ## Bug fixes
 
 * Fixed issue in `model_info()` with `stan_polr()` models.